Move files to new folder location
