+++ projects/_03_independent_cqc/_02_clean/tests/utils/ascwds_filled_posts_calculator/test_total_staff_equals_worker_records.py
@@ -1,27 +1,25 @@
 import unittest
 import warnings
 
-from tests.test_file_data import CalculateAscwdsFilledPostsDifferenceInRangeData as Data
+import projects._03_independent_cqc._02_clean.utils.ascwds_filled_posts_calculator.total_staff_equals_worker_records as job
+from tests.test_file_data import (
+    CalculateAscwdsFilledPostsTotalStaffEqualWorkerRecordsData as Data,
+)
 from tests.test_file_schemas import (
-    CalculateAscwdsFilledPostsDifferenceInRangeSchemas as Schemas,
+    CalculateAscwdsFilledPostsTotalStaffEqualWorkerRecordsSchemas as Schemas,
 )
-
 from utils import utils
-from utils.column_names.ind_cqc_pipeline_columns import (
-    IndCqcColumns as IndCQC,
-)
-
-import utils.ind_cqc_filled_posts_utils.ascwds_filled_posts_calculator.calculate_ascwds_filled_posts_difference_within_range as job
+from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 
-class TestAscwdsFilledPostsAbsoluteDiffInRange(unittest.TestCase):
+class TestAscwdsFilledPostsWorkerRecordsEqualsTotalStaff(unittest.TestCase):
     def setUp(self):
         self.spark = utils.get_spark()
         self.ascwds_total_staff_and_worker_record_df = self.spark.createDataFrame(
             Data.calculate_ascwds_filled_posts_rows,
             Schemas.calculate_ascwds_filled_posts_schema,
         )
-        self.df = job.calculate_ascwds_filled_posts_difference_within_range(
+        self.df = job.calculate_ascwds_filled_posts_totalstaff_equal_wkrrecs(
             self.ascwds_total_staff_and_worker_record_df,
             IndCQC.total_staff_bounded,
             IndCQC.worker_records_bounded,
@@ -35,7 +33,7 @@ class TestAscwdsFilledPostsAbsoluteDiffInRange(unittest.TestCase):
     ):
         self.assertEqual(self.df.count(), 9)
 
-    def test_calculate_ascwds_filled_posts_difference_within_range_values(
+    def test_calculate_ascwds_filled_posts_totalstaff_equal_wkrrecs_values(
         self,
     ):
         df = self.df.sort(IndCQC.location_id).collect()
@@ -45,29 +43,26 @@ class TestAscwdsFilledPostsAbsoluteDiffInRange(unittest.TestCase):
         self.assertEqual(df[3][IndCQC.ascwds_filled_posts], None)
         self.assertEqual(df[4][IndCQC.ascwds_filled_posts], None)
         self.assertEqual(df[5][IndCQC.ascwds_filled_posts], None)
-        self.assertEqual(df[6][IndCQC.ascwds_filled_posts], 11.5)
-        self.assertEqual(df[7][IndCQC.ascwds_filled_posts], 487.5)
+        self.assertEqual(df[6][IndCQC.ascwds_filled_posts], None)
+        self.assertEqual(df[7][IndCQC.ascwds_filled_posts], None)
         self.assertEqual(df[8][IndCQC.ascwds_filled_posts], 8.0)
 
-    def test_calculate_ascwds_filled_posts_difference_within_range_sources(
+    def test_calculate_ascwds_filled_posts_totalstaff_equal_wkrrecs_sources(
         self,
     ):
         df = self.df.sort(IndCQC.location_id).collect()
         self.assertEqual(df[0][IndCQC.ascwds_filled_posts_source], None)
         self.assertEqual(
             df[1][IndCQC.ascwds_filled_posts_source],
-            job.ascwds_filled_posts_difference_within_range_source_description,
+            job.ascwds_filled_posts_totalstaff_equal_wkrrecs_source_description,
         )
         self.assertEqual(df[2][IndCQC.ascwds_filled_posts_source], None)
         self.assertEqual(df[3][IndCQC.ascwds_filled_posts_source], None)
         self.assertEqual(df[4][IndCQC.ascwds_filled_posts_source], None)
         self.assertEqual(df[5][IndCQC.ascwds_filled_posts_source], None)
-        self.assertEqual(
-            df[6][IndCQC.ascwds_filled_posts_source],
-            job.ascwds_filled_posts_difference_within_range_source_description,
-        )
+        self.assertEqual(df[6][IndCQC.ascwds_filled_posts_source], None)
         self.assertEqual(
             df[7][IndCQC.ascwds_filled_posts_source],
-            job.ascwds_filled_posts_difference_within_range_source_description,
+            None,
         )
         self.assertEqual(df[8][IndCQC.ascwds_filled_posts_source], "already populated")
