+++ projects/_03_independent_cqc/_02_clean/tests/utils/ascwds_filled_posts_calculator/test_utils.py
@@ -10,9 +10,9 @@ from pyspark.sql.types import (
     DoubleType,
 )
 
+import projects._03_independent_cqc._02_clean.utils.ascwds_filled_posts_calculator.utils as job
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
-import utils.ind_cqc_filled_posts_utils.ascwds_filled_posts_calculator.utils as job
 from tests.test_file_data import (
     CalculateAscwdsFilledPostsUtilsData as Data,
 )
