+++ projects/_03_independent_cqc/_02_clean/utils/ascwds_filled_posts_calculator/utils.py
@@ -1,6 +1,6 @@
 from pyspark.sql import DataFrame, functions as F
 
-from utils.ind_cqc_filled_posts_utils.ascwds_filled_posts_calculator.calculation_constants import (
+from projects._03_independent_cqc._02_clean.utils.ascwds_filled_posts_calculator.calculation_constants import (
     ASCWDSFilledPostCalculationConstants as calculation_constant,
 )
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
